Add files via upload
Based on Rashid's workshop on systematic literature analysis, the prompts and collected data have been reorganized. 

Now we can customize specific questions for filtering the literature (yes or no questions) or to set some very detailed specific questions (such as what the research question of the article is).

I have got time to do the translation yet. Will be updated later if needed.